requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-APScheduler
Flask-Cors
Werkzeug
Pillow
PyMuPDF
python-docx
python-pptx
matplotlib
openpyxl
pandas
chardet